implement 'work setup' implied project endpoint; update help docs; update deps
