+++ package.json
@@ -34,7 +34,7 @@
   "devDependencies": {
     "@liquid-labs/catalyst-scripts": "^1.0.0-alpha.57",
     "@liquid-labs/federated-json": "^1.0.0-alpha.31",
-    "@liquid-labs/git-toolkit": "^1.0.0-alpha.9",
+    "@liquid-labs/git-toolkit": "^1.0.0-alpha.10",
     "@liquid-labs/http-smart-response": "^1.0.0-alpha.3",
     "@liquid-labs/liq-handlers-lib": "^1.0.0-alpha.12",
     "@liquid-labs/versioning": "^1.0.0-alpha.1",
@@ -42,10 +42,10 @@
     "natural-sort": "^1.0.0"
   },
   "dependencies": {
-    "@liquid-labs/github-toolkit": "^1.0.0-alpha.5",
+    "@liquid-labs/github-toolkit": "^1.0.0-alpha.6",
     "@liquid-labs/liq-credentials-db": "^1.0.0-alpha.3",
     "@liquid-labs/liq-projects-lib": "^1.0.0-alpha.3",
-    "@liquid-labs/liq-qa-lib": "file:.yalc/@liquid-labs/liq-qa-lib",
+    "@liquid-labs/liq-qa-lib": "^1.0.0-alpha.1",
     "@liquid-labs/octocache": "^1.0.0-alpha.3",
     "http-errors": "^2.0.0",
     "shelljs": "^0.8.5"
