+++ src/handlers/projects/setup-implied.mjs
@@ -2,11 +2,11 @@ import createError from 'http-errors'
 
 import { determineImpliedProject } from '@liquid-labs/liq-projects-lib'
 
-import { doRename, getRenameEndpointParameters } from './_lib/rename-lib'
+import { doSetup, getSetupEndpointParameters } from './_lib/setup-lib'
 
-const path = ['projects', 'rename']
+const path = ['projects', 'setup']
 
-const { help, method, parameters } = getRenameEndpointParameters({ workDesc : 'implied' })
+const { help, method, parameters } = getSetupEndpointParameters({ workDesc : 'implied' })
 
 const func = ({ model, reporter }) => async(req, res) => {
   reporter = reporter.isolate()
@@ -17,7 +17,7 @@ const func = ({ model, reporter }) => async(req, res) => {
   }
   const [orgKey, localProjectName] = determineImpliedProject({ currDir : cwd }).split('/')
 
-  await doRename({ localProjectName, model, orgKey, reporter, req, res })
+  await doSetup({ localProjectName, model, orgKey, reporter, req, res })
 }
 
 export {
